requirements.txt
# API
fastapi
uvicorn
python-multipart

# Core
numpy
pandas
pillow

# Image Processing
opencv-python
albumentations

# OCR
pytesseract
easyocr

# NLP / Transformers
spacy
transformers
torch
torchvision
sentencepiece

# Utilities
regex

pdf2image
